requirements.txt
jupyter==1.0.0
jupyterlab==4.0.9
ipykernel==6.26.0
pytesseract==0.3.10
Pillow==10.1.0
openai==1.51.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.3
openpyxl==3.1.2
